requirements.txt
streamlit>=1.37
pandas>=2.0
gspread>=6.0
google-auth>=2.30
google-auth-oauthlib>=1.2
requests>=2.32
